Fix building docs
