+++ .github/workflows/build_docs.yml
@@ -1,76 +1,52 @@
 # Simple workflow for deploying static content to GitHub Pages
-name: Deploy static content to Pages
+name: Build docs
 
-on: [push]
+on:
+  pull_request:
+    branches:
+      - "main"
 
-
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-permissions:
-  contents: read
-  pages: write
-  id-token: write
-
-# Allow one concurrent deployment
-concurrency:
-  group: "pages"
-  cancel-in-progress: true
-
-env:
-  # Directory that will be published on github pages
-  PUBLISH_DIR: ./_build/html
+  # This workflow can be executed inside another workfow file
+  workflow_call:
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
 
 jobs:
-
   build:
     runs-on: ubuntu-22.04
-    container: ghcr.io/scientificcomputing/fenics-gmsh:2024-02-19
+    container: ghcr.io/scientificcomputing/fenics-gmsh:2023-08-16
+    env:
+      DEB_PYTHON_INSTALL_LAYOUT: deb_system
+      PUBLISH_DIR: ./_build/html
+      DISPLAY: ":99.0"
+      PYVISTA_TRAME_SERVER_PROXY_PREFIX: "/proxy/"
+      PYVISTA_TRAME_SERVER_PROXY_ENABLED: "True"
+      PYVISTA_OFF_SCREEN: True
+      PYVISTA_JUPYTER_BACKEND: "panel"
 
     steps:
       - name: Checkout
         uses: actions/checkout@v4
 
-      - name: Cache
-        id: cache
-        uses: actions/cache@v3
-        with:
-          path: |
-            ~/.cache/instant
-            ~/.cache/dijitso
-            ~/.cache/pip
-            ~/_build
-          key: cache_v1
-          restore-keys: |
-            cache_v1
+      - name: "Install pyvista dependencies"
+        run: apt-get update && apt-get install -y libgl1-mesa-glx libxrender1 xvfb nodejs
 
       - name: Install dependencies
-        run: python3 -m pip install ".[docs]"
+        run: python3 -m pip install ".[docs,examples,pyvista]"
+
+      - name: Cache notebooks
+        id: cache-notebooks
+        uses: actions/cache@v4
+        with:
+          path: _build
+          key: cache_key_1 # Bump number manually when you want to clear the cache
 
       - name: Build docs
         run: jupyter book build .
 
-      - name: Upload artifact
-        uses: actions/upload-pages-artifact@v2
+      - name: Upload documentation as artifact
+        uses: actions/upload-artifact@v4
         with:
+          name: documentation
           path: ${{ env.PUBLISH_DIR }}
-
-  # Single deploy job since we're just deploying
-  deploy:
-    if: github.ref == 'refs/heads/main'
-    needs: build
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Setup Pages
-        uses: actions/configure-pages@v3
-
-
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: actions/deploy-pages@v2
+          if-no-files-found: error
