+++ _config.yml
@@ -43,4 +43,4 @@ sphinx:
   - 'sphinx.ext.napoleon'
   - 'sphinx.ext.viewcode'
 
-exclude_patterns: [".pytest_cache/*"]
+exclude_patterns: [".pytest_cache/*" ,"tests/.pytest_cache/*"]
