+++ _toc.yml
@@ -11,9 +11,8 @@ parts:
     - file: demos/simple_ode.py
     - file: demos/solve_many_odes.py
     - file: demos/niederer_benchmark.py
-    - file: demos/planar_slab.py
+    - file: demos/slab.py
     - file: demos/diffusion.py
-    - file: demos/endocardial_stimulation_single.py
     - file: demos/endocardial_stimulation.py
   - caption: Verification
     chapters:
