+++ demos/niederer_benchmark.py
@@ -1,13 +1,13 @@
 # # Niederer benchmark
-#
 # In this example we will use the same setup as in the Niederer benchmark
 # > Niederer SA, Kerfoot E, Benson AP, Bernabeu MO, Bernus O, Bradley C,
-#  Cherry EM, Clayton R, Fenton FH, Garny A, Heidenreich E, Land S, Maleckar M,
-#  Pathmanathan P, Plank G, Rodríguez JF, Roy I, Sachse FB, Seemann G, Skavhaug O,
-#  Smith NP. Verification of cardiac tissue electrophysiology simulators using an
-#  N-version benchmark. Philos Trans A Math Phys Eng Sci. 2011 Nov 13;369(1954):4331-51.
-#  doi: 10.1098/rsta.2011.0139. PMID: 21969679; PMCID: PMC3263775.
+#   Cherry EM, Clayton R, Fenton FH, Garny A, Heidenreich E, Land S, Maleckar M,
+#   Pathmanathan P, Plank G, Rodríguez JF, Roy I, Sachse FB, Seemann G, Skavhaug O,
+#   Smith NP. Verification of cardiac tissue electrophysiology simulators using an
+#   N-version benchmark. Philos Trans A Math Phys Eng Sci. 2011 Nov 13;369(1954):4331-51.
+#   doi: 10.1098/rsta.2011.0139. PMID: 21969679; PMCID: PMC3263775.
 #
+
 from pathlib import Path
 import json
